+++ CPP03/ex01/ScavTrap.cpp
@@ -101,7 +101,7 @@ void ScavTrap::takeDamage(unsigned int amount)
 
 void ScavTrap::beRepaired(unsigned int amount)
 {
-	if (this->hp + amount <= this->max_hp)
+	if (this->hp + (int)amount <= this->max_hp)
 		this->hp += amount;
 	else
 		this->hp = this->max_hp;
